+++ docs/dukey.txt
@@ -1,3 +1,3 @@
-1.[T][ ] kill akshay 
-2.[T][ ] kill minjun 
-3.[D][ ] kill (by: 9)
+1.[T][ ] whats up g 
+2.[T][ ] nthuin fam 
+3.[T][ ] cmon
