+++ src/main/java/Parser.java
@@ -1,0 +1,11 @@
+import java.util.regex.*;
+
+public class Parser {
+    public String parseCommandInput(String line) {
+        String[] words = line.split(" ", 2);
+        return words[0];
+    }
+
+
+}
+
