+++ src/main/java/Ui.java
@@ -2,7 +2,7 @@ import Tasks.*;
 import java.util.ArrayList;
 
 public class Ui {
-    public static void showWelcomeMessage() {
+    public void showWelcomeMessage() {
         System.out.println("Hey! I'm Dukey, your virtual assistant!\nWhat can I do for you?\n");
     }
 
@@ -10,8 +10,18 @@ public class Ui {
         System.out.println("Bye. Hope to see you again soon!\n");
     }
 
+    public static void outputHeader() {
+        printLine();
+        System.out.println("Here are the matching tasks in your list:\n");
+    }
 
+    public static void printLine() {
+        System.out.println("_____________________________________________________");
+    }
 
+    public static void showLoadingError() {
+        System.out.println("Error loading tasks from text file to Dukey");
+    }
     public void showTaskList(ArrayList<Task> tasks) {
         System.out.println("_____________________________________________________");
         System.out.println("Here are the tasks in your list:");
